+++ CHANGELOG.md
@@ -42,3 +42,4 @@
 - Gene Attribute Modifiers are kept on respawn
 - When you have Max Health 1 or 2, you'll respawn with completely full health rather than just the vanilla 10 hearts
 - Lowered default cooldown for Dragon's Breath from 15 seconds to 5
+- Improved messages for Genes on cooldown